Make ServerModule class into a method
